backend/requirements.txt
fastapi 
uvicorn
httpx
sqlalchemy
python-dotenv 
google-generativeai 
numpy